agent/requirements.txt
coinbase-agentkit==0.2.3
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.7
python-dotenv==1.1.1
